erp_django/requirements.txt
django==2.1.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
pytest==3.7.4
pytest-django==3.4.2
requests==2.19.1
PyDrive==1.3.1
yagmail==0.11.2014
python-google-calendar-api==0.5
celery==4.2.1
redis==2.10.6
django-guardian==1.4.9
djangorestframework-guardian==0.1.0